--- include/math_functions.c
@@ -1,6 +1,0 @@
-#include "math_functions.h"
-
-int fsign(float comp) 
-{
-	return ((0) < comp) - (comp < (0));
-}
