--- include/math_functions.h
@@ -1,2 +1,0 @@
-// returns -1, 0, or 1, depending if the inputed value is negative, zero, or positive
-int fsign(float comp);
