--- include/nes.h
@@ -1,52 +1,0 @@
-typedef struct NES
-{
-	Uint32 color;
-} NES;
-
-NES NES_PAL00 = {0x6A6A6AFF};
-NES NES_PAL01 = {0x001380FF};
-NES NES_PAL02 = {0x1E008AFF};
-NES NES_PAL03 = {0x39007AFF};
-NES NES_PAL04 = {0x550056FF};
-NES NES_PAL05 = {0x5A0018FF};
-NES NES_PAL06 = {0x4F1000FF};
-NES NES_PAL07 = {0x3D1C00FF};
-NES NES_PAL08 = {0x253200FF};
-NES NES_PAL09 = {0x003D00FF};
-NES NES_PAL0A = {0x004000FF};
-NES NES_PAL0B = {0x003924FF};
-NES NES_PAL0C = {0x002E55FF};
-NES NES_PAL0D = {0x000000FF};
-NES NES_PAL0E = {0x000000FF};
-NES NES_PAL0F = {0x00000000};	// TRANSPARENCY
-NES NES_PALINVALID = {0xFF00FFFF};
-
-NES NES_PAL20 = {0xFFFFFFFF};
-/*
-namespace NES
-{
-	static const Uint32
-		// RED, GREEN, BLUE, ALPHA
-		PAL00{0x6A6A6AFF}, PAL01{0x001380FF}, PAL02{0x1E008AFF}, PAL03{0x39007AFF},
-		PAL04{0x550056FF}, PAL05{0x5A0018FF}, PAL06{0x4F1000FF}, PAL07{0x3D1C00FF}, 
-		PAL08{0x253200FF}, PAL09{0x003D00FF}, PAL0A{0x004000FF}, PAL0B{0x003924FF},
-		PAL0C{0x002E55FF}, PAL0D{0x000000FF}, PAL0E{0x000000FF}, PAL0F{0x00000000},
-
-		PAL10{0xB9BCB9FF}, PAL11{0x1850C7FF}, PAL12{0x4B30E3FF}, PAL13{0x9B22D6FF},
-		PAL14{0x951FA9FF}, PAL15{0x9D285CFF}, PAL16{0x983700FF}, PAL17{0x7F4C00FF}, 
-		PAL18{0x5E6400FF}, PAL19{0x227700FF}, PAL1A{0x027E02FF}, PAL1B{0x007645FF}, 
-		PAL1C{0x006E8AFF}, PAL1D{0x000000FF}, PAL1E{0x000000FF}, PAL1F{0x00000000},
-
-		PAL20{0xFFFFFFFF}, PAL21{0x68A6FFFF}, PAL22{0x8C9CFFFF}, PAL23{0xB586F4FF},
-		PAL24{0xD975FDFF}, PAL25{0xE377B9FF}, PAL26{0xE58D68FF}, PAL27{0xD49D29FF},
-		PAL28{0xB3AF0CFF}, PAL29{0x7BC211FF}, PAL2A{0x55CA47FF}, PAL2B{0x46CB81FF},
-		PAL2C{0x47C1C5FF}, PAL2D{0x4A4D4AFF}, PAL2E{0x000000FF}, PAL2F{0x00000000},
-
-		PAL30{0xFFFFFFFF}, PAL31{0xCCEAFFFF}, PAL32{0xDDDEFFFF}, PAL33{0xECDAFFFF},
-		PAL34{0xF8D7FEFF}, PAL35{0xFCD6D5FF}, PAL36{0xFDDBCFFF}, PAL37{0xF9E7B5FF},
-		PAL38{0xF1F0AAFF}, PAL39{0xDAFAA9FF}, PAL3A{0xC9FFBCFF}, PAL3B{0xC3FBD7FF},
-		PAL3C{0xC4F6F6FF}, PAL3D{0xBEC1BEFF}, PAL3E{0x000000FF}, PAL3F{0x00000000};
-		
-}
-
-*/
