+++ src/main.c
@@ -1,39 +1,31 @@
+// put your libs here.
 #include "Raquet.h"
+// TODO: WRITE FPS FUNCTIONS
+// TODO: COMMENT MORE STUFF IN THE Raquet.h FILE
+// TODO: MAKE CREATING VARIABLES MORE SIMPLE
 
-// Math and Graphical LIBS
-#include "math_functions.c"
-#include "ppf_functions.c"
-
-// GAME LIBS
-#include "input.h"
-
-// CONSTANTS
+uint8_t demoy;
+uint8_t demox;
+NES palface[3];
 SDL_Texture* placeface;
 
-// GAME VARS
-SDL_Color backgroundColorControl = {255, 0, 255, 255};
+void createthedog()
+{
+	LoadPPFData("./assets/main.ppf");
 
-uint8_t demox = 0;
-uint8_t demoy = 0;
+	 palface[0] = NES_PAL00;
+	 palface[1] = NES_PAL0D;
+	 palface[2] = NES_PAL20;
+	
+	placeface = LoadCHR(0, palface);
 
-int initRaquet()
-{	
-	int ready = 1;
-	if (!initsdl())
-	{
-		printf("Failed to Initialize SDL\n");
-		ready = 0;
-	}
-	// PUT YOUR SETUP CODE HERE
-	LoadPPFData("./assets/test.ppf");
+	demox = 0;
+	demoy = 0;
 	
-	
-	return ready;
 }
 
-int runthedog()
+void runthedog()
 {
-
 	int move = key_right - key_left;
 	int vmove = key_down - key_up;
 	demox += move;
@@ -42,12 +34,8 @@ int runthedog()
 	// Draw our stuffs
 	RaquetClear(NES_PALINVALID);
 
-	if (placeface == NULL) {
-		NES paltest[3] = {NES_PAL0D, NES_PAL00, NES_PAL20};
-		placeface = LoadCHR(0, paltest);
-	}
-
-	for (int y = 0; y < SCREEN_WIDTH/8; y++) {
+	for (int y = 0; y < SCREEN_WIDTH/8; y++) 
+	{
 		for (int x = 0; x < SCREEN_WIDTH/8; x++)
 		{
 			PlaceCHR(placeface, demox + (x * 8), demoy + (y * 8));
@@ -61,40 +49,11 @@ int runthedog()
 	// Reset the Window
 	RaquetUpdate();
 
-	return 1;
 }
 
 int main() {
 
-	if (!initRaquet())
-	{
-		printf("Failed to Initialize\n");
-	}
-	else
-	{
-		// hack 
-		int quit = 0; 
-		while(quit == 0)
-		{ 
-			while(SDL_PollEvent(&e))
-			{
-				handleInput(e);
-				if(e.type == SDL_QUIT)
-				{
-					quit = 1;
-				}
-				
-			}
-
-			// do our game stuff
-			runthedog();
-			
-		}
-			
-	}
-	
-	quitit();
-
+	RaquetMain();
 	return 0;
 	
 }
